mcp/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
redis==4.5.4
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
pydantic==1.10.7
python-multipart==0.0.6
celery==5.2.7
openai==0.27.4
anthropic==0.3.11
requests==2.28.2
python-dotenv==1.0.0
pypdf==3.9.0